+++ packages/const/package.json
@@ -6,7 +6,8 @@
     ".": "./src/index.ts",
     "./currency": "./src/currency.ts",
     "./desktopGlobalShortcuts": "./src/desktopGlobalShortcuts.ts",
-    "./hotkeys": "./src/hotkeys.ts"
+    "./hotkeys": "./src/hotkeys.ts",
+    "./visualRef": "./src/visualRef.ts"
   },
   "main": "./src/index.ts",
   "dependencies": {
