+++ packages/prompts/src/prompts/files/index.test.ts
@@ -1,5 +1,5 @@
+import { createVisualFileRef } from '@lobechat/const/visualRef';
 import type { ChatFileItem, ChatImageItem, ChatVideoItem } from '@lobechat/types';
-import { createVisualFileRef } from '@lobechat/types';
 import { describe, expect, it } from 'vitest';
 
 import { filesPrompts } from './index';
