+++ packages/types/src/tool/index.ts
@@ -31,4 +31,3 @@ export * from './manifest';
 export * from './plugin';
 export * from './search';
 export * from './tool';
-export * from './visualRef';
