+++ src/server/services/toolExecution/serverRuntimes/__tests__/lobeAgent.test.ts
@@ -1,5 +1,6 @@
 import { LobeAgentIdentifier, MAX_VISUAL_MEDIA_URLS } from '@lobechat/builtin-tool-lobe-agent';
-import { createVisualFileRef, RequestTrigger } from '@lobechat/types';
+import { createVisualFileRef } from '@lobechat/const/visualRef';
+import { RequestTrigger } from '@lobechat/types';
 import { beforeEach, describe, expect, it, vi } from 'vitest';
 
 import type { ToolExecutionContext } from '../../types';
